refactor(clean-arch): break analysis module dependency cycle
HeartbeatCallback is imported from _config (its definition site) instead of
via the subprocess re-export, which removes the _types -> subprocess edge and
collapses the documented _command/_mcp_config cycle. The two deferred-import
shims become plain module-level imports of cache.local.default_cache_root,
and AnalysisConfig.run_config regains its real RunConfig type.
